Changes from SteamBroker
I think if should work correct in this way. Still have a problem with Aspect ratio. You make this with and without option in different places and it's to complicated to get into all this. I think you should remove "without aspect and Intrinsics fuction" becouse this is not needed, Also you should divide Center_x, center_y shift using fo center of distortion and intrinsincs is used to move and stretch image before distortion function.